requirements.txt
firebase_admin==6.6.0
matplotlib==3.10.0
python-dotenv==1.0.1
Requests==2.32.3
ttkthemes==3.2.2